requirements.txt
discord.py
openai-whisper
discord-ext-voice-recv
pynacl